requirements.txt
gensim==3.8.3
matplotlib==3.3.4
mysql-connector-python==8.0.25
networkx==2.5.1
nltk==3.6.2
numpy==1.19.5
pandas==1.1.5
requests==2.25.1
scikit-learn==0.24.2
scikit-network==0.22.0
scipy==1.5.4
six==1.16.0
sklearn==0.0
tagme==0.1.3
